test: refactor runner
