+++ crates/node_binding/rspack.wasi.cjs
@@ -108,62 +108,4 @@ const { instance: __napiInstance, module: __wasiModule, napiModule: __napiModule
   },
 })
 module.exports = __napiModule.exports
-module.exports.Assets = __napiModule.exports.Assets
-module.exports.AsyncDependenciesBlock = __napiModule.exports.AsyncDependenciesBlock
-module.exports.Chunk = __napiModule.exports.Chunk
-module.exports.ChunkGraph = __napiModule.exports.ChunkGraph
-module.exports.ChunkGroup = __napiModule.exports.ChunkGroup
-module.exports.Chunks = __napiModule.exports.Chunks
-module.exports.CodeGenerationResult = __napiModule.exports.CodeGenerationResult
-module.exports.CodeGenerationResults = __napiModule.exports.CodeGenerationResults
-module.exports.ConcatenatedModule = __napiModule.exports.ConcatenatedModule
-module.exports.ContextModule = __napiModule.exports.ContextModule
-module.exports.Dependency = __napiModule.exports.Dependency
-module.exports.Diagnostics = __napiModule.exports.Diagnostics
-module.exports.EntryDataDto = __napiModule.exports.EntryDataDto
-module.exports.EntryDataDTO = __napiModule.exports.EntryDataDTO
-module.exports.EntryDependency = __napiModule.exports.EntryDependency
-module.exports.EntryOptionsDto = __napiModule.exports.EntryOptionsDto
-module.exports.EntryOptionsDTO = __napiModule.exports.EntryOptionsDTO
-module.exports.ExternalModule = __napiModule.exports.ExternalModule
-module.exports.JsCompilation = __napiModule.exports.JsCompilation
-module.exports.JsCompiler = __napiModule.exports.JsCompiler
-module.exports.JsContextModuleFactoryAfterResolveData = __napiModule.exports.JsContextModuleFactoryAfterResolveData
-module.exports.JsContextModuleFactoryBeforeResolveData = __napiModule.exports.JsContextModuleFactoryBeforeResolveData
-module.exports.JsDependencies = __napiModule.exports.JsDependencies
-module.exports.JsEntries = __napiModule.exports.JsEntries
-module.exports.JsExportsInfo = __napiModule.exports.JsExportsInfo
-module.exports.JsModuleGraph = __napiModule.exports.JsModuleGraph
-module.exports.JsResolver = __napiModule.exports.JsResolver
-module.exports.JsResolverFactory = __napiModule.exports.JsResolverFactory
-module.exports.JsStats = __napiModule.exports.JsStats
-module.exports.KnownBuildInfo = __napiModule.exports.KnownBuildInfo
-module.exports.Module = __napiModule.exports.Module
-module.exports.ModuleGraphConnection = __napiModule.exports.ModuleGraphConnection
-module.exports.NativeWatcher = __napiModule.exports.NativeWatcher
-module.exports.NativeWatchResult = __napiModule.exports.NativeWatchResult
-module.exports.NormalModule = __napiModule.exports.NormalModule
-module.exports.RawExternalItemFnCtx = __napiModule.exports.RawExternalItemFnCtx
-module.exports.ReadonlyResourceData = __napiModule.exports.ReadonlyResourceData
-module.exports.ResolverFactory = __napiModule.exports.ResolverFactory
-module.exports.Sources = __napiModule.exports.Sources
-module.exports.VirtualFileStore = __napiModule.exports.VirtualFileStore
-module.exports.JsVirtualFileStore = __napiModule.exports.JsVirtualFileStore
-module.exports.async = __napiModule.exports.async
-module.exports.BuiltinPluginName = __napiModule.exports.BuiltinPluginName
-module.exports.cleanupGlobalTrace = __napiModule.exports.cleanupGlobalTrace
-module.exports.EnforceExtension = __napiModule.exports.EnforceExtension
-module.exports.EXPECTED_RSPACK_CORE_VERSION = __napiModule.exports.EXPECTED_RSPACK_CORE_VERSION
-module.exports.formatDiagnostic = __napiModule.exports.formatDiagnostic
-module.exports.JsLoaderState = __napiModule.exports.JsLoaderState
-module.exports.JsRspackSeverity = __napiModule.exports.JsRspackSeverity
-module.exports.loadBrowserslist = __napiModule.exports.loadBrowserslist
-module.exports.minify = __napiModule.exports.minify
-module.exports.minifySync = __napiModule.exports.minifySync
-module.exports.RawRuleSetConditionType = __napiModule.exports.RawRuleSetConditionType
-module.exports.registerGlobalTrace = __napiModule.exports.registerGlobalTrace
-module.exports.RegisterJsTapKind = __napiModule.exports.RegisterJsTapKind
-module.exports.sync = __napiModule.exports.sync
-module.exports.syncTraceEvent = __napiModule.exports.syncTraceEvent
-module.exports.transform = __napiModule.exports.transform
-module.exports.transformSync = __napiModule.exports.transformSync
+
