+++ packages/rspack-test-tools/src/runner/index.ts
@@ -1,9 +1,7 @@
 export * from "./basic";
 export * from "./cache";
 export * from "./hot";
-export * from "./hot-step";
 export * from "./multiple";
-export * from "./normal";
 export * from "./runner";
 export * from "./type";
 export * from "./watch";
