+++ packages/rspack-test-tools/src/runner/runner/basic.ts
@@ -1,5 +1,8 @@
 import fs from "node:fs";
 import path from "node:path";
+import { fileURLToPath, pathToFileURL } from "node:url";
+import vm, { SourceTextModule } from "node:vm";
+import asModule from "../../helper/legacy/asModule";
 import urlToRelativePath from "../../helper/legacy/urlToRelativePath";
 import type {
 	ECompilerType,
@@ -9,13 +12,18 @@ import type {
 	TCompilerStatsCompilation,
 	TTestConfig
 } from "../../type";
-import type {
-	IBasicGlobalContext,
-	IBasicModuleScope,
-	TBasicRunnerFile,
-	TModuleObject,
-	TRunnerRequirer
+import {
+	EEsmMode,
+	type IBasicGlobalContext,
+	type IBasicModuleScope,
+	type TBasicRunnerFile,
+	type TModuleObject,
+	type TRunnerRequirer
 } from "../type";
+
+declare global {
+	var printLogger: boolean;
+}
 
 const isRelativePath = (p: string) => /^\.\.?\//.test(p);
 const getSubPath = (p: string) => {
@@ -52,11 +60,11 @@ export interface IBasicRunnerOptions<T extends ECompilerType> {
 	compilerOptions: TCompilerOptions<T>;
 	cachable?: boolean;
 }
-
-export abstract class BasicRunner<
-	T extends ECompilerType = ECompilerType.Rspack
-> implements ITestRunner
+export class BasicRunner<T extends ECompilerType = ECompilerType.Rspack>
+	implements ITestRunner
 {
+	protected requireCache = Object.create(null);
+
 	protected globalContext: IBasicGlobalContext | null = null;
 	protected baseModuleScope: IBasicModuleScope | null = null;
 	protected requirers: Map<string, TRunnerRequirer> = new Map();
@@ -101,13 +109,100 @@ export abstract class BasicRunner<
 		return ((this.globalContext || {}) as Record<string, unknown>)[name];
 	}
 
-	protected abstract createGlobalContext(): IBasicGlobalContext;
-	protected abstract createBaseModuleScope(): IBasicModuleScope;
-	protected abstract createModuleScope(
+	protected createGlobalContext(): IBasicGlobalContext {
+		return {
+			console: {
+				log: (...args: any[]) => {
+					if (printLogger) {
+						console.log(...args);
+					}
+				},
+				warn: (...args: any[]) => {
+					if (printLogger) {
+						console.warn(...args);
+					}
+				},
+				error: (...args: any[]) => {
+					console.error(...args);
+				},
+				info: (...args: any[]) => {
+					if (printLogger) {
+						console.info(...args);
+					}
+				},
+				debug: (...args: any[]) => {
+					if (printLogger) {
+						console.info(...args);
+					}
+				},
+				trace: (...args: any[]) => {
+					if (printLogger) {
+						console.info(...args);
+					}
+				},
+				assert: (...args: any[]) => {
+					console.assert(...args);
+				},
+				clear: () => {
+					console.clear();
+				}
+			},
+			setTimeout: ((
+				cb: (...args: any[]) => void,
+				ms: number | undefined,
+				...args: any
+			) => {
+				const timeout = setTimeout(cb, ms, ...args);
+				timeout.unref();
+				return timeout;
+			}) as typeof setTimeout,
+			clearTimeout: clearTimeout
+		};
+	}
+	protected createBaseModuleScope(): IBasicModuleScope {
+		const baseModuleScope: IBasicModuleScope = {
+			console: this.globalContext!.console,
+			setTimeout: this.globalContext!.setTimeout,
+			clearTimeout: this.globalContext!.clearTimeout,
+			nsObj: (m: Object) => {
+				Object.defineProperty(m, Symbol.toStringTag, {
+					value: "Module"
+				});
+				return m;
+			},
+			process,
+			URL,
+			Blob,
+			Symbol,
+			Buffer,
+			setImmediate,
+			__MODE__: this._options.compilerOptions.mode,
+			__SNAPSHOT__: path.join(this._options.source, "__snapshot__"),
+			...this._options.env
+		};
+		return baseModuleScope;
+	}
+	protected createModuleScope(
 		requireFn: TRunnerRequirer,
 		m: TModuleObject,
 		file: TBasicRunnerFile
-	): IBasicModuleScope;
+	): IBasicModuleScope {
+		const requirer: TRunnerRequirer & {
+			webpackTestSuiteRequire?: boolean;
+		} = requireFn.bind(null, path.dirname(file.path));
+		requirer.webpackTestSuiteRequire = true;
+		return {
+			...this.baseModuleScope!,
+			require: requirer,
+			module: m,
+			exports: m.exports,
+			__dirname: path.dirname(file.path),
+			__filename: file.path,
+			_globalAssign: {
+				expect: this._options.env.expect
+			}
+		};
+	}
 
 	protected getFile(
 		modulePath: string[] | string,
@@ -160,11 +255,208 @@ export abstract class BasicRunner<
 	protected postExecute(m: Object, file: TBasicRunnerFile) {}
 
 	protected createRunner() {
-		this.requirers.set(
-			"entry",
-			(currentDirectory, modulePath, context = {}) => {
-				throw new Error("Not implement");
+		this.requirers.set("cjs", this.createCjsRequirer());
+		this.requirers.set("esm", this.createEsmRequirer());
+		this.requirers.set("miss", this.createMissRequirer());
+		this.requirers.set("json", this.createJsonRequirer());
+		this.requirers.set("entry", (currentDirectory, modulePath, context) => {
+			const file = this.getFile(modulePath, currentDirectory);
+			if (!file) {
+				return this.requirers.get("miss")!(currentDirectory, modulePath);
 			}
-		);
+			if (file.path.endsWith(".json")) {
+				return this.requirers.get("json")!(currentDirectory, modulePath, {
+					...context,
+					file
+				});
+			}
+			if (
+				file.path.endsWith(".mjs") &&
+				this._options.compilerOptions.experiments?.outputModule
+			) {
+				return this.requirers.get("esm")!(currentDirectory, modulePath, {
+					...context,
+					file
+				});
+			}
+			return this.requirers.get("cjs")!(currentDirectory, modulePath, {
+				...context,
+				file
+			});
+		});
+	}
+
+	protected createMissRequirer(): TRunnerRequirer {
+		return (currentDirectory, modulePath, context = {}) => {
+			const modulePathStr = modulePath as string;
+			const modules = this._options.testConfig.modules;
+			if (modules && modulePathStr in modules) {
+				return modules[modulePathStr];
+			}
+			return require(
+				modulePathStr.startsWith("node:")
+					? modulePathStr.slice(5)
+					: modulePathStr
+			);
+		};
+	}
+
+	protected createJsonRequirer(): TRunnerRequirer {
+		return (currentDirectory, modulePath, context = {}) => {
+			if (Array.isArray(modulePath)) {
+				throw new Error("Array module path is not supported in hot cases");
+			}
+			const file = context.file || this.getFile(modulePath, currentDirectory);
+			if (!file) {
+				return this.requirers.get("miss")!(currentDirectory, modulePath);
+			}
+			return JSON.parse(
+				fs.readFileSync(path.join(this._options.dist, modulePath), "utf-8")
+			);
+		};
+	}
+
+	protected createCjsRequirer(): TRunnerRequirer {
+		return (currentDirectory, modulePath, context = {}) => {
+			if (modulePath === "@rspack/test-tools") {
+				return require("@rspack/test-tools");
+			}
+			const file = context.file || this.getFile(modulePath, currentDirectory);
+			if (!file) {
+				return this.requirers.get("miss")!(currentDirectory, modulePath);
+			}
+
+			if (file.path in this.requireCache) {
+				return this.requireCache[file.path].exports;
+			}
+
+			const m = {
+				exports: {},
+				webpackTestSuiteModule: true
+			};
+			this.requireCache[file.path] = m;
+			const currentModuleScope = this.createModuleScope(
+				this.getRequire(),
+				m,
+				file
+			);
+
+			if (this._options.testConfig.moduleScope) {
+				this._options.testConfig.moduleScope(
+					currentModuleScope,
+					this._options.stats,
+					this._options.compilerOptions
+				);
+			}
+
+			if (!this._options.runInNewContext) {
+				file.content = `Object.assign(global, _globalAssign);\n ${file.content}`;
+			}
+			if (file.content.includes("__STATS__") && this._options.stats) {
+				currentModuleScope.__STATS__ = this._options.stats();
+			}
+			if (file.content.includes("__STATS_I__")) {
+				const statsIndex = this._options.stats?.()?.__index__;
+				if (typeof statsIndex === "number") {
+					currentModuleScope.__STATS_I__ = statsIndex;
+				}
+			}
+			const args = Object.keys(currentModuleScope);
+			const argValues = args.map(arg => currentModuleScope[arg]);
+			const code = `(function(${args.join(", ")}) {
+        ${file.content}
+      })`;
+
+			this.preExecute(code, file);
+			const fn = this._options.runInNewContext
+				? vm.runInNewContext(code, this.globalContext!, file.path)
+				: vm.runInThisContext(code, file.path);
+			fn.call(
+				this._options.testConfig.nonEsmThis
+					? this._options.testConfig.nonEsmThis(modulePath)
+					: m.exports,
+				...argValues
+			);
+
+			this.postExecute(m, file);
+			return m.exports;
+		};
+	}
+
+	protected createEsmRequirer(): TRunnerRequirer {
+		const esmContext = vm.createContext(this.baseModuleScope!, {
+			name: "context for esm"
+		});
+		const esmCache = new Map<string, SourceTextModule>();
+		const esmIdentifier = this._options.name;
+
+		return (currentDirectory, modulePath, context = {}) => {
+			if (!SourceTextModule) {
+				throw new Error(
+					"Running this test requires '--experimental-vm-modules'.\nRun with 'node --experimental-vm-modules node_modules/jest-cli/bin/jest'."
+				);
+			}
+			const _require = this.getRequire();
+			const file = context.file || this.getFile(modulePath, currentDirectory);
+			if (!file) {
+				return this.requirers.get("miss")!(currentDirectory, modulePath);
+			}
+
+			if (file.content.includes("__STATS__")) {
+				esmContext.__STATS__ = this._options.stats?.();
+			}
+
+			let esm = esmCache.get(file.path);
+			if (!esm) {
+				esm = new SourceTextModule(file.content, {
+					identifier: `${esmIdentifier}-${file.path}`,
+					// no attribute
+					url: `${pathToFileURL(file.path).href}?${esmIdentifier}`,
+					context: esmContext,
+					initializeImportMeta: (meta: { url: string }, _: any) => {
+						meta.url = pathToFileURL(file!.path).href;
+					},
+					importModuleDynamically: async (
+						specifier: any,
+						module: { context: any }
+					) => {
+						const result = await _require(path.dirname(file!.path), specifier, {
+							esmMode: EEsmMode.Evaluated
+						});
+						return await asModule(result, module.context);
+					}
+				} as any);
+				esmCache.set(file.path, esm);
+			}
+			if (context.esmMode === EEsmMode.Unlinked) return esm;
+			return (async () => {
+				await esm.link(async (specifier, referencingModule) => {
+					return await asModule(
+						await _require(
+							path.dirname(
+								referencingModule.identifier
+									? referencingModule.identifier.slice(esmIdentifier.length + 1)
+									: fileURLToPath((referencingModule as any).url)
+							),
+							specifier,
+							{
+								esmMode: EEsmMode.Unlinked
+							}
+						),
+						referencingModule.context,
+						true
+					);
+				});
+				if ((esm as any).instantiate) (esm as any).instantiate();
+				await esm.evaluate();
+				if (context.esmMode === EEsmMode.Evaluated) {
+					return esm;
+				}
+				const ns = esm.namespace as {
+					default: unknown;
+				};
+				return ns.default && ns.default instanceof Promise ? ns.default : ns;
+			})();
+		};
 	}
 }
