+++ packages/rspack-test-tools/src/runner/runner/index.ts
@@ -1,8 +1,3 @@
 export * from "./basic";
-export * from "./cjs";
-export * from "./esm";
-export * from "./normal";
 export * from "./watch";
 export * from "./web";
-export * from "./web/fake";
-export * from "./web/jsdom";
