+++ packages/rspack-test-tools/src/runner/runner/web/index.ts
@@ -2,11 +2,10 @@ import {
 	type ECompilerType,
 	EDocumentType,
 	type ITestRunner
-} from "../../type";
-import type { BasicRunner, IBasicRunnerOptions } from "./basic";
-import type { CommonJsRunner } from "./cjs";
-import { FakeDocumentWebRunner } from "./web/fake";
-import { JSDOMWebRunner } from "./web/jsdom";
+} from "../../../type";
+import type { BasicRunner, IBasicRunnerOptions } from "../basic";
+import { FakeDocumentWebRunner } from "./fake";
+import { JSDOMWebRunner } from "./jsdom";
 
 export interface IWebRunnerOptions<
 	T extends ECompilerType = ECompilerType.Rspack
@@ -17,7 +16,7 @@ export interface IWebRunnerOptions<
 export class WebRunner<T extends ECompilerType = ECompilerType.Rspack>
 	implements ITestRunner
 {
-	protected originMethods: Partial<CommonJsRunner> = {};
+	protected originMethods: Partial<BasicRunner<T>> = {};
 	private implement: BasicRunner<T>;
 	constructor(protected _webOptions: IWebRunnerOptions<T>) {
 		const { dom } = _webOptions;
